Version 0.9.1.0dev2:
Rank menu logic changed:
  - only input file is required
  - input format is a multisampled MAF file
  - profile, cohort_size and observed mutations would be extracted from input file
  - the file is actually processed twice: once for nucleic acid mutations and once for protein mutations, the overhead is small
  - if cohort is requested -c will give a list of available precalculated Mutagene cohorts (based on COSMIC data)
  - if cohort is specified it will override profile, cohort_size and observed mutations
  - it is possible to override profile and cohort_size separately

Only the first encountered transcript for a gene is considered. Mutations in other transcripts would be ignored in the analysis. This is important to avoid overcounting of observed mutations in MAF files with per-transcript annotation of mutations.

Currently pseudo-counts are no longer added to observed mutations data even if precalculated cohort is used. This changes results compared to previous implementation.